Starting a fresh GameMaker Studio project
Based on a Udemy Course by Benjamin Anderson (aka HEARTBEAST) which details how to make a simple 2D side-scroller in GameMaker version. 1.4. Updating to modern (2024 and beyond) GameMaker in order to learn the ropes of game development. Starting from the top and attempting some semblance of proper version control and software engineering principles and other things I got a four-year degree for but never applied professionally for some reason.