more project cleanup
Confirmed that the game still runs and we can play!

That said, the score and high score are still bugged tf out

There is literally no reason for the nested folder nightmare we're currently doing. Also we have a redundant .gitattributes file that we're overriding with whatever GameMaker Studio auto-generated.